@@ -2,16 +2,16 @@ package de.maxhenkel.better_respawn;
 
 import de.maxhenkel.better_respawn.capabilities.RespawnPosition;
 import de.maxhenkel.better_respawn.capabilities.SpawnPointCapabilityProvider;
-import net.minecraft.entity.Entity;
-import net.minecraft.entity.player.PlayerEntity;
-import net.minecraft.entity.player.ServerPlayerEntity;
-import net.minecraft.entity.player.SpawnLocationHelper;
-import net.minecraft.util.ResourceLocation;
-import net.minecraft.util.math.BlockPos;
-import net.minecraft.util.math.ChunkPos;
-import net.minecraft.util.math.vector.Vector3d;
-import net.minecraft.util.math.vector.Vector3i;
-import net.minecraft.world.server.ServerWorld;
+import net.minecraft.core.BlockPos;
+import net.minecraft.core.Vec3i;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.server.level.PlayerRespawnLogic;
+import net.minecraft.server.level.ServerLevel;
+import net.minecraft.server.level.ServerPlayer;
+import net.minecraft.world.entity.Entity;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.level.ChunkPos;
+import net.minecraft.world.phys.Vec3;
 import net.minecraftforge.common.util.FakePlayer;
 import net.minecraftforge.event.AttachCapabilitiesEvent;
 import net.minecraftforge.event.entity.EntityJoinWorldEvent;
@@ -29,8 +29,7 @@ public class RespawnEvents {
     private static final int FIND_SPAWN_ATTEMPTS = 16;
     private static final ResourceLocation RESPAWN_CAPABILITY_ID = new ResourceLocation(Main.MODID, "respawn_location");
 
-    private Random random;
-
+    private final Random random;
 
     public RespawnEvents() {
         random = new Random();
@@ -38,13 +37,10 @@ public class RespawnEvents {
 
     @SubscribeEvent
     public void onDeath(LivingDeathEvent event) {
-        if (!(event.getEntity() instanceof ServerPlayerEntity)) {
+        if (!(event.getEntity() instanceof ServerPlayer player)) {
             return;
         }
 
-        ServerPlayerEntity player = (ServerPlayerEntity) event.getEntity();
-
-        // canRespawnHere
         if (!player.getLevel().dimensionType().bedWorks()) {
             return;
         }
@@ -52,11 +48,11 @@ public class RespawnEvents {
         BlockPos bedLocation = player.getRespawnPosition();
 
         if (bedLocation != null) {
-            Optional<Vector3d> vec3d = PlayerEntity.findRespawnPositionAndUseSpawnBlock(player.getLevel(), bedLocation, 0F, false, false);
+            Optional<Vec3> vec3d = Player.findRespawnPositionAndUseSpawnBlock(player.getLevel(), bedLocation, 0F, false, false);
             if (vec3d.isPresent()) {
-                Vector3d spawn = vec3d.get();
-                if (player.blockPosition().distManhattan(new Vector3i(spawn.x, spawn.y, spawn.z)) <= Main.SERVER_CONFIG.bedRange.get()) {
-                    Main.LOGGER.debug("Player {} is within the range of its bed", player.getName().getContents());
+                Vec3 spawn = vec3d.get();
+                if (player.blockPosition().distManhattan(new Vec3i(spawn.x, spawn.y, spawn.z)) <= Main.SERVER_CONFIG.bedRange.get()) {
+                    Main.LOGGER.info("Player {} is within the range of its respawn block", player.getName().getContents());
                     return;
                 }
             }
@@ -69,15 +65,14 @@ public class RespawnEvents {
         }
 
         player.setRespawnPosition(player.level.dimension(), respawnPos, 0F, true, false);
-        Main.LOGGER.debug("Set temporary respawn location to [{}, {}, {}]", respawnPos.getX(), respawnPos.getY(), respawnPos.getZ());
+        Main.LOGGER.info("Set temporary respawn location to [{}, {}, {}]", respawnPos.getX(), respawnPos.getY(), respawnPos.getZ());
     }
 
     @SubscribeEvent
     public void onJoinWorld(EntityJoinWorldEvent event) {
-        if (!(event.getEntity() instanceof ServerPlayerEntity) || !event.getEntity().isAlive()) {
+        if (!(event.getEntity() instanceof ServerPlayer player) || !event.getEntity().isAlive()) {
             return;
         }
-        ServerPlayerEntity player = (ServerPlayerEntity) event.getEntity();
         RespawnPosition respawnPosition = player.getCapability(Main.RESPAWN_CAPABILITY).orElse(null);
 
         if (respawnPosition == null) {
@@ -87,9 +82,9 @@ public class RespawnEvents {
         BlockPos respawn = respawnPosition.getPos(event.getWorld());
         player.setRespawnPosition(event.getWorld().dimension(), respawn, 0F, false, false);
         if (respawn == null) {
-            Main.LOGGER.debug("Setting the players respawn position back to world spawn");
+            Main.LOGGER.info("Setting the players respawn position back to world spawn");
         } else {
-            Main.LOGGER.debug("Setting the players respawn position back to [{}, {}, {}]", respawn.getX(), respawn.getY(), respawn.getZ());
+            Main.LOGGER.info("Setting the players respawn position back to [{}, {}, {}]", respawn.getX(), respawn.getY(), respawn.getZ());
         }
     }
 
@@ -101,7 +96,7 @@ public class RespawnEvents {
         BlockPos newSpawn = event.getNewSpawn();
         if (newSpawn != null) {
             event.getPlayer().getCapability(Main.RESPAWN_CAPABILITY).ifPresent(respawnPosition -> respawnPosition.setPos(event.getPlayer().level, newSpawn));
-            Main.LOGGER.debug("Updating the respawn location of player {} to [{}, {}, {}]", event.getPlayer().getName().getContents(), newSpawn.getX(), newSpawn.getY(), newSpawn.getZ());
+            Main.LOGGER.info("Updating the respawn location of player {} to [{}, {}, {}]", event.getPlayer().getName().getContents(), newSpawn.getX(), newSpawn.getY(), newSpawn.getZ());
         }
     }
 
@@ -110,38 +105,40 @@ public class RespawnEvents {
         if (!event.isWasDeath()) {
             return;
         }
+        event.getOriginal().reviveCaps();
         event.getOriginal().getCapability(Main.RESPAWN_CAPABILITY).ifPresent(respawnPosition -> {
             event.getPlayer().getCapability(Main.RESPAWN_CAPABILITY).ifPresent(respawnPosition1 -> {
                 respawnPosition1.copyFrom(respawnPosition);
-                Main.LOGGER.debug("Copying respawn location capability of player {}", event.getPlayer().getName().getContents());
+                Main.LOGGER.info("Copying respawn location capability of player {}", event.getPlayer().getName().getContents());
             });
         });
+        event.getOriginal().invalidateCaps();
     }
 
     @SubscribeEvent
     public void onPlayerCapabilityAttach(AttachCapabilitiesEvent<Entity> event) {
-        if (event.getObject() instanceof PlayerEntity && !(event.getObject() instanceof FakePlayer)) {
+        if (event.getObject() instanceof Player && !(event.getObject() instanceof FakePlayer)) {
             if (!event.getObject().getCapability(Main.RESPAWN_CAPABILITY).isPresent()) {
                 event.addCapability(RESPAWN_CAPABILITY_ID, new SpawnPointCapabilityProvider());
-                Main.LOGGER.debug("Attaching respawn capability to player");
+                Main.LOGGER.info("Attaching respawn capability to player");
             }
         }
     }
 
     @Nullable
-    public BlockPos findValidRespawnLocation(ServerWorld world, BlockPos deathLocation) {
+    public BlockPos findValidRespawnLocation(ServerLevel world, BlockPos deathLocation) {
         int min = Main.SERVER_CONFIG.minRespawnDistance.get();
         int max = Main.SERVER_CONFIG.maxRespawnDistance.get();
 
         BlockPos pos = null;
         for (int i = 0; i < FIND_SPAWN_ATTEMPTS && pos == null; i++) {
-            Main.LOGGER.debug("Searching for respawn location - Attempt {}/{}", i + 1, FIND_SPAWN_ATTEMPTS);
-            pos = SpawnLocationHelper.getSpawnPosInChunk(world, new ChunkPos(new BlockPos(getRandomRange(deathLocation.getX(), min, max), 0, getRandomRange(deathLocation.getZ(), min, max))), true);
+            Main.LOGGER.info("Searching for respawn location - Attempt {}/{}", i + 1, FIND_SPAWN_ATTEMPTS);
+            pos = PlayerRespawnLogic.getSpawnPosInChunk(world, new ChunkPos(new BlockPos(getRandomRange(deathLocation.getX(), min, max), 0, getRandomRange(deathLocation.getZ(), min, max))), true);
         }
         if (pos == null) {
-            Main.LOGGER.debug("Found no valid respawn location after {} attempts", FIND_SPAWN_ATTEMPTS);
+            Main.LOGGER.info("Found no valid respawn location after {} attempts", FIND_SPAWN_ATTEMPTS);
         } else {
-            Main.LOGGER.debug("Found valid respawn location: [{}, {}, {}]", pos.getX(), pos.getY(), pos.getZ());
+            Main.LOGGER.info("Found valid respawn location: [{}, {}, {}]", pos.getX(), pos.getY(), pos.getZ());
         }
         return pos;
     }
