@@ -1,19 +1,17 @@
 package de.maxhenkel.better_respawn;
 
 import de.maxhenkel.better_respawn.capabilities.RespawnPosition;
-import de.maxhenkel.better_respawn.capabilities.SpawnPointCapabilityStorage;
 import de.maxhenkel.better_respawn.net.MessageRespawnDelay;
 import de.maxhenkel.corelib.CommonRegistry;
 import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.common.capabilities.Capability;
 import net.minecraftforge.common.capabilities.CapabilityInject;
-import net.minecraftforge.common.capabilities.CapabilityManager;
-import net.minecraftforge.eventbus.api.SubscribeEvent;
+import net.minecraftforge.common.capabilities.RegisterCapabilitiesEvent;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.config.ModConfig;
 import net.minecraftforge.fml.event.lifecycle.FMLCommonSetupEvent;
 import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
-import net.minecraftforge.fml.network.simple.SimpleChannel;
+import net.minecraftforge.fmllegacy.network.simple.SimpleChannel;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
@@ -33,20 +31,22 @@ public class Main {
 
     public Main() {
         FMLJavaModLoadingContext.get().getModEventBus().addListener(this::commonSetup);
+        FMLJavaModLoadingContext.get().getModEventBus().addListener(this::onRegisterCapabilities);
 
         SERVER_CONFIG = CommonRegistry.registerConfig(ModConfig.Type.SERVER, ServerConfig.class);
     }
 
-    @SubscribeEvent
     public void commonSetup(FMLCommonSetupEvent event) {
         MinecraftForge.EVENT_BUS.register(this);
         MinecraftForge.EVENT_BUS.register(new RespawnEvents());
         MinecraftForge.EVENT_BUS.register(new RespawnTimerEvents());
 
-        CapabilityManager.INSTANCE.register(RespawnPosition.class, new SpawnPointCapabilityStorage(), () -> new RespawnPosition());
-
         SIMPLE_CHANNEL = CommonRegistry.registerChannel(Main.MODID, "default");
         CommonRegistry.registerMessage(SIMPLE_CHANNEL, 0, MessageRespawnDelay.class);
+    }
+
+    public void onRegisterCapabilities(RegisterCapabilitiesEvent event) {
+        event.register(RespawnPosition.class);
     }
 
 }
